// repo/server/src/common/enums/audit-action.enum.ts
export enum AuditAction {
  USER_CREATED = 'USER_CREATED',
  USER_UPDATED = 'USER_UPDATED',
  USER_DEACTIVATED = 'USER_DEACTIVATED',
  USER_ACTIVATED = 'USER_ACTIVATED',
  USER_PASSWORD_RESET = 'USER_PASSWORD_RESET',
  PR_CREATED = 'PR_CREATED',
  PR_SUBMITTED = 'PR_SUBMITTED',
  PR_AUTO_APPROVED = 'PR_AUTO_APPROVED',
  PR_APPROVED = 'PR_APPROVED',
  PR_REJECTED = 'PR_REJECTED',
  PR_CANCELLED = 'PR_CANCELLED',
  PO_CREATED = 'PO_CREATED',
  PO_ISSUED = 'PO_ISSUED',
  PO_CANCELLED = 'PO_CANCELLED',
  PO_STATUS_CHANGED = 'PO_STATUS_CHANGED',
  RECEIPT_CREATED = 'RECEIPT_CREATED',
  RECEIPT_COMPLETED = 'RECEIPT_COMPLETED',
  RA_CREATED = 'RA_CREATED',
  RA_SUBMITTED = 'RA_SUBMITTED',
  RA_STATUS_CHANGED = 'RA_STATUS_CHANGED',
  RA_COMPLETED = 'RA_COMPLETED',
  ARTICLE_CREATED = 'ARTICLE_CREATED',
  ARTICLE_UPDATED = 'ARTICLE_UPDATED',
  ARTICLE_PROMOTED = 'ARTICLE_PROMOTED',
  ARTICLE_ARCHIVED = 'ARTICLE_ARCHIVED',
}