// repo/server/src/common/enums/notification-type.enum.ts
export enum NotificationType {
  REQUEST_APPROVED = 'REQUEST_APPROVED',
  REQUEST_REJECTED = 'REQUEST_REJECTED',
  PO_ISSUED = 'PO_ISSUED',
  RECEIPT_COMPLETED = 'RECEIPT_COMPLETED',
  RETURN_CREATED = 'RETURN_CREATED',
  ARTICLE_PUBLISHED = 'ARTICLE_PUBLISHED',
  SYSTEM_ALERT = 'SYSTEM_ALERT',
  SCHEDULE_CHANGE = 'SCHEDULE_CHANGE',
  CANCELLATION = 'CANCELLATION',
  REVIEW_OUTCOME = 'REVIEW_OUTCOME',
}